+++ modules/server/src/main/ts/BedrockAuto.ts
@@ -6,11 +6,14 @@ import * as RunnerRoutes from './bedrock/server/RunnerRoutes';
 import * as Reporter from './bedrock/core/Reporter';
 import * as DriverMaster from './bedrock/server/DriverMaster';
 import * as Driver from './bedrock/auto/Driver';
+import * as Tunnel from './bedrock/auto/Tunnel';
 import * as Lifecycle from './bedrock/core/Lifecycle';
 import { BedrockAutoSettings } from './bedrock/core/Settings';
 import { ExitCodes } from './bedrock/util/ExitCodes';
 import * as ConsoleReporter from './bedrock/core/ConsoleReporter';
 import * as SettingsResolver from './bedrock/core/SettingsResolver';
+import * as portfinder from 'portfinder';
+import { format } from 'node:util';
 
 export const go = (bedrockAutoSettings: BedrockAutoSettings): void => {
   console.log('bedrock-auto ' + Version.get() + ' starting...');
@@ -21,11 +24,36 @@ export const go = (bedrockAutoSettings: BedrockAutoSettings): void => {
   const isPhantom = browserName === 'phantomjs';
   const isHeadless = settings.browser.endsWith('-headless') || isPhantom;
   const basePage = 'src/resources/html/' + (isPhantom ? 'bedrock-phantom.html' : 'bedrock.html');
+  // Remote settings
+  const remoteWebdriver = settings.remote;
+  const sishDomain = settings.sishDomain;
+  const username = settings.username ?? process.env.LT_USERNAME;
+  const accesskey = settings.accesskey ?? process.env.LT_ACCESS_KEY;
 
   const routes = RunnerRoutes.generate('auto', settings.projectdir, settings.basedir, settings.config, settings.bundler, settings.testfiles, settings.chunk, settings.retries, settings.singleTimeout, settings.stopOnFailure, basePage, settings.coverage, settings.polyfills);
 
   routes.then(async (runner) => {
+    const shutdownServices: (() => Promise<any>)[] = [];
 
+    // LambdaTest Tunnel must know dev server port, but tunnel must be created before dev server.
+    const servicePort = await portfinder.getPortPromise({
+      port: 8000,
+      stopPort: 20000
+    });
+
+    const tunnelCredentials = {
+      user: username,
+      key: accesskey
+    };
+
+    const tunnel = await Tunnel.prepareConnection(servicePort, remoteWebdriver, sishDomain, tunnelCredentials);
+    shutdownServices.push(tunnel.shutdown);
+    const location = tunnel.url.href;
+
+    console.log('Creating webdriver...');
+    if (remoteWebdriver == 'aws') {
+        console.log('INFO: Webdriver creation waits for device farm session to activate. Takes 30-45s.');
+    }
     const driver = await Driver.create({
       browser: browserName,
       basedir: settings.basedir,
@@ -35,8 +63,13 @@ export const go = (bedrockAutoSettings: BedrockAutoSettings): void => {
       extraBrowserCapabilities: settings.extraBrowserCapabilities,
       verbose: settings.verbose,
       wipeBrowserCache: settings.wipeBrowserCache,
+      remoteWebdriver,
       webdriverPort: settings.webdriverPort,
-      useSelenium: settings.useSelenium
+      useSelenium: settings.useSelenium,
+      username,
+      accesskey,
+      devicefarmRegion: settings.devicefarmRegion,
+      deviceFarmArn: settings.devicefarmArn
     });
 
     const webdriver = driver.webdriver;
@@ -45,17 +78,20 @@ export const go = (bedrockAutoSettings: BedrockAutoSettings): void => {
       driver: Attempt.passed(webdriver),
       master,
       runner,
-      stickyFirstSession: true
+      stickyFirstSession: true,
+      port: servicePort
     });
+    shutdownServices.push(service.shutdown, driver.shutdown);
 
-    const shutdown = () => Promise.all([ service.shutdown(), driver.shutdown() ]);
+    const shutdown = () => Promise.allSettled(shutdownServices.map((shutdown_fn) => shutdown_fn()));
 
     try {
       if (!isHeadless) {
-        console.log('bedrock-auto ' + Version.get() + ' available at: http://localhost:' + service.port);
+        console.log('bedrock-auto ' + Version.get() + ' available at: ' + location);
       }
 
-      await webdriver.url('http://localhost:' + service.port);
+      console.log('Loading initial page...');
+      await webdriver.url(location);
       console.log(isPhantom ? '\nPhantom tests loading ...\n' : '\nInitial page has loaded ...\n');
       service.markLoaded();
       service.enableHud();
@@ -73,7 +109,8 @@ export const go = (bedrockAutoSettings: BedrockAutoSettings): void => {
       return Lifecycle.error(e, webdriver, shutdown, settings.gruntDone, settings.delayExit);
     }
   }).catch((err) => {
-    console.error(chalk.red(err));
+    // Chalk does not use a formatter. Using node's built-in to expand Objects, etc.
+    console.error(chalk.red('Error creating webdriver', format(err)));
     Lifecycle.exit(settings.gruntDone, ExitCodes.failures.unexpected);
   });
 };