+++ modules/server/src/main/ts/bedrock/cli/Clis.ts
@@ -49,7 +49,13 @@ export const forAuto = (directories: Directories, argv: string[] = process.argv)
     ClOptions.extraBrowserCapabilities,
     ClOptions.skipResetMousePosition,
     ClOptions.wipeBrowserCache,
+    ClOptions.remote,
     ClOptions.webdriverPort,
+    ClOptions.sishDomain,
+    ClOptions.username,
+    ClOptions.accesskey,
+    ClOptions.devicefarmArn,
+    ClOptions.devicefarmRegion,
     ClOptions.useSelenium
   ]), argv) as Attempt<cli.CliError, BedrockAutoSettings>;
 };