+++ modules/server/src/test/ts/ClisTest.ts
@@ -60,8 +60,9 @@ describe('Clis.forAuto', () => {
       useSelenium: false,
       extraBrowserCapabilities: '',
       skipResetMousePosition: false,
-      webdriverPort: 4444,
-      wipeBrowserCache: false
+      wipeBrowserCache: false,
+      remote: '',
+      webdriverPort: 4444
     }, cleanResult(actual));
   });
 
@@ -99,8 +100,9 @@ describe('Clis.forAuto', () => {
       useSelenium: false,
       extraBrowserCapabilities: '',
       skipResetMousePosition: false,
-      webdriverPort: 4444,
-      wipeBrowserCache: false
+      wipeBrowserCache: false,
+      remote: '',
+      webdriverPort: 4444
     }, cleanResult(actual));
   });
 
@@ -139,8 +141,9 @@ describe('Clis.forAuto', () => {
       useSelenium: false,
       extraBrowserCapabilities: '',
       skipResetMousePosition: false,
-      webdriverPort: 4444,
-      wipeBrowserCache: false
+      wipeBrowserCache: false,
+      remote: '',
+      webdriverPort: 4444
     }, cleanResult(actual));
   });
 
@@ -178,8 +181,97 @@ describe('Clis.forAuto', () => {
       useSelenium: false,
       extraBrowserCapabilities: ' --some-browser-flag',
       skipResetMousePosition: false,
-      webdriverPort: 4444,
-      wipeBrowserCache: false
+      wipeBrowserCache: false,
+      remote: '',
+      webdriverPort: 4444
+    }, cleanResult(actual));
+  });
+
+  it('parses values for remote testing using lambdatest', () => {
+    const args = [
+      '--browser', 'chrome-headless',
+      '--files', 'src/test/resources/test.file1',
+      '--username', 'lt_username',
+      '--accesskey', 'lt_access_key',
+      '--remote', 'lambdatest'
+    ];
+    const actual = Clis.forAuto(directories, args);
+    AttemptUtils.assertResult({
+      browser: 'chrome-headless',
+      bundler: 'webpack',
+      config: 'tsconfig.json',
+      name: 'bedrock-run',
+      output: 'scratch',
+      help: false,
+      testfiles: [
+        'src/test/resources/test.file1'
+      ],
+      delayExit: false,
+      singleTimeout: 30000,
+      stopOnFailure: false,
+      overallTimeout: 600000,
+      loglevel: 'advanced',
+      version: false,
+      chunk: 100,
+      retries: 0,
+      polyfills: [ 'Symbol' ],
+      verbose: false,
+      bucket: 1,
+      buckets: 1,
+      useSandboxForHeadless: false,
+      useSelenium: false,
+      extraBrowserCapabilities: '',
+      skipResetMousePosition: false,
+      wipeBrowserCache: false,
+      remote: 'lambdatest',
+      username: 'lt_username',
+      accesskey: 'lt_access_key',
+      webdriverPort: 4444
+    }, cleanResult(actual));
+  });
+
+  it('parses values for remote testing using aws devicefarm', () => {
+    const args = [
+      '--browser', 'chrome-headless',
+      '--files', 'src/test/resources/test.file1',
+      '--sishDomain', 'sish.osu.tiny.work',
+      '--devicefarmRegion', 'us-west-2',
+      '--devicefarmArn', 'arn:aws:devicefarm:us-west-2:123123123123:project-name:123a123b-123c-123d-1234a-123a123b123c',
+      '--remote', 'aws'
+    ];
+    const actual = Clis.forAuto(directories, args);
+    AttemptUtils.assertResult({
+      browser: 'chrome-headless',
+      bundler: 'webpack',
+      config: 'tsconfig.json',
+      name: 'bedrock-run',
+      output: 'scratch',
+      help: false,
+      testfiles: [
+        'src/test/resources/test.file1'
+      ],
+      delayExit: false,
+      singleTimeout: 30000,
+      stopOnFailure: false,
+      overallTimeout: 600000,
+      loglevel: 'advanced',
+      version: false,
+      chunk: 100,
+      retries: 0,
+      polyfills: [ 'Symbol' ],
+      verbose: false,
+      bucket: 1,
+      buckets: 1,
+      useSandboxForHeadless: false,
+      useSelenium: false,
+      extraBrowserCapabilities: '',
+      skipResetMousePosition: false,
+      wipeBrowserCache: false,
+      remote: 'aws',
+      sishDomain: 'sish.osu.tiny.work',
+      devicefarmRegion: 'us-west-2',
+      devicefarmArn: 'arn:aws:devicefarm:us-west-2:123123123123:project-name:123a123b-123c-123d-1234a-123a123b123c',
+      webdriverPort: 4444
     }, cleanResult(actual));
   });